update type tests
