+++ test/drivers/github.test.ts
@@ -18,7 +18,7 @@ describe("drivers: github", () => {
   });
 
   it("can read a json file content", async () => {
-    const pkg = (await storage.getItem("package.json")) as Record<
+    const pkg = (await storage.getItem("package.json"))! as Record<
       string,
       unknown
     >;
